kenya weather/requirements.txt
# requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
apache-airflow==2.7.1
python-dotenv==1.0.0